Couple of fixes for older Qt versions. Patch from Tony Lill on -dev.
git-svn-id: http://svn.mythtv.org/svn/branches/release-0-18-fixes@6078 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/NuppelVideoPlayer.cpp
@@ -3744,7 +3744,8 @@ int NuppelVideoPlayer::FlagCommercials(bool showPercentage, bool fullSpeed,
                 {
                     JobQueue::ChangeJobComment(jobID, QObject::tr(
                                                "%1 Frames Completed @ %2 fps.")
-                                               .arg(currentFrame->frameNumber)
+                                               .arg((long)
+                                                    currentFrame->frameNumber)
                                                .arg(flagFPS));
                 }
             }
